separate submit transaction methods into their own test files
